+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make module available via "Web > Info > Changelog information" instead
   of own main backend module.
 - Change extension icon
+- Use michelf/php-markdown instead of fluidtypo3/vhs for rendering
 
 ## [1.1.1] - 2024-12-16